+++ src/features/post/ui/PostDetailDialog.tsx
