+++ src/pages/PostsManagerPage.tsx
@@ -2,11 +2,11 @@ import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from
 import { useEffect, useState } from "react"
 import { useLocation, useNavigate } from "react-router-dom"
 
-import CommentAddDialog from "@/features/ui/comment/CommentAddDialog"
-import CommentEditDialog from "@/features/ui/comment/CommentEditDialog"
-import PostAddCardHeader from "@/features/ui/post/PostAddCardHeader"
-import PostDetailDialog from "@/features/ui/post/PostDetailDialog"
-import PostEditDialog from "@/features/ui/post/PostEditDialog"
+import CommentAddDialog from "@/features/comment/ui/CommentAddDialog"
+import CommentEditDialog from "@/features/comment/ui/CommentEditDialog"
+import PostAddCardHeader from "@/features/post/ui/PostAddCardHeader"
+import PostDetailDialog from "@/features/post/ui/PostDetailDialog"
+import PostEditDialog from "@/features/post/ui/PostEditDialog"
 import { Button } from "../shared/ui/Button"
 import { Card, CardContent } from "../shared/ui/card"
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../shared/ui/dialog"
